+++ src/projectCreator.ts
@@ -1,6 +1,0 @@
-
-import { Downloader } from './downloader';
-
-export class ProjectCreator {
-    
-}