+++ src/ProjectMaker/projectMaker.ts
@@ -1,0 +1,9 @@
+"use strict";
+
+import { Uri } from "vscode";
+
+
+
+export class ProjectMaker {
+    
+}