+++ src/diskFunctions.ts
@@ -1,7 +1,8 @@
+import { BlobOptions } from "buffer";
 import * as fs from "fs";
 
 export class DiskFunctions {
-    static readFromFile(file: string): String|null {
+    static readFromFile(file: string): String | null {
         try {
             return fs.readFileSync(file).toString();
         } catch (err) {
@@ -23,8 +24,43 @@ export class DiskFunctions {
             return false;
         }
     }
-    static writeToFile(file: string, content: string): Boolean {
+    /**
+     * Extract a path from a path to a file
+     * @param path path to a file (must include a directory / or \\ )
+     * @returns if path not found an empty string is returned
+     */
+    static getDirectoryFromFilePath(path: string): string {
+        let pos = path.lastIndexOf("/");
+        if (pos < 0) {
+            pos = path.lastIndexOf("\\");
+        }
+        if (pos < 0) {
+            return "";
+        }
+
+        return path.substring(0, pos);
+    }
+    static createDirectory(dir: string, createPathRecursively: boolean = false): Boolean {
         try {
+            if (!DiskFunctions.dirExists(dir)) {
+                return fs.mkdirSync(dir, { recursive: createPathRecursively }) !== undefined;
+            }
+        } catch (err) {
+            return false;
+        }
+        return true;
+    }
+    static writeToFile(file: string, content: string, createPath: boolean = false): Boolean {
+        try {
+            if (createPath) {
+                const dir = this.getDirectoryFromFilePath(file);
+                if (dir.length < 1) {
+                    return false;
+                }
+                if (!this.createDirectory(dir, true)) {
+                    return false;
+                }
+            }
             fs.writeFileSync(file, content);
             return true;
         } catch (err) {