+++ src/tools/preparePublish.js
@@ -37,7 +37,30 @@ var mdTableEntry = (key, snippet) => {
     return `| ${prefix} | ${key} | ${snippet.description} |\n`;
 };
 
-var readSnippetsToMdTableRows = (dir) => {
+var getSnippetsFromFile = (filePath, inSnippets) => {
+    var snippets = inSnippets? inSnippets : [];
+    let stat = fs.statSync(filePath);
+
+    if (stat.isFile()) {
+        console.log("Reading '%s' ", filePath);
+        snippets.push(JSON.parse(fs.readFileSync(filePath)));
+    }
+
+    return snippets;
+};
+
+var readSnippetsInFileToMdTableRows = (filePath) => {
+    var snippets = getSnippetsFromFile(filePath);
+    var mdContent = "";
+    snippets.forEach(fileContent => {
+        Object.keys(fileContent).forEach(function (key) {
+            mdContent += mdTableEntry(key, fileContent[key]);
+        });
+    });
+    return mdContent;
+};
+
+var readSnippetsInDirToMdTableRows = (dir) => {
 
     var filenames = fs.readdirSync(dir);
     var snippets = [];
@@ -45,12 +68,9 @@ var readSnippetsToMdTableRows = (dir) => {
     filenames.forEach(function (file, index) {
         // Make one pass and make the file complete
         var filePath = path.join(dir, file);
-        let stat = fs.statSync(filePath);
 
-        if (stat.isFile()) {
-            console.log("Reading '%s' ", filePath);
-            snippets.push(JSON.parse(fs.readFileSync(filePath)));
-        }
+        snippets = getSnippetsFromFile(filePath, snippets);
+
     });
 
     snippets.forEach(fileContent => {
@@ -62,31 +82,30 @@ var readSnippetsToMdTableRows = (dir) => {
 };
 
 var buildAndReplaceContent = () => {
-    var cMdCppTableRows = readSnippetsToMdTableRows(`${snippetDirs}${slash}cpp`);
-    var cMdCMakeTableRows = readSnippetsToMdTableRows(`${snippetDirs}${slash}cmake`);
+    var cMdCppTableRows = readSnippetsInDirToMdTableRows(`${snippetDirs}${slash}cpp`);
+    var cMdCMakeTableRows = readSnippetsInDirToMdTableRows(`${snippetDirs}${slash}cmake`);
+    var cMdGitIgnoreTableRows = readSnippetsInFileToMdTableRows(`${snippetDirs}${slash}gitignore.json`);
 
 
     console.log(`Exporting snippet list to markdown file ${mdFile}`);
     if (lib.fileExists(mdFile)) {
         console.log('File exists');
     }
-    var mdContent = lib.replaceContent(fs.readFileSync(mdFile).toString(),
-        "|:--------|:------|:------------|\n",
-        "\n### CMake snippets",
-        cMdCppTableRows
-    );
-    var ret = lib.replaceContent(mdContent,
-        "|:-------|:------|:------------|\n",
-        "\n\n[Top](#gepper-readme)",
-        cMdCMakeTableRows
-    );
-    return ret;
+    var mdContent;
+    mdContent = lib.replaceContent(fs.readFileSync(mdFile).toString(), "|:--------|:------|:------------|\n", "\n### CMake snippets", cMdCppTableRows);
+    mdContent = lib.replaceContent(mdContent, "|:-------|:------|:------------|\n", "\n### .gitignore snippets", cMdCMakeTableRows);
+    mdContent = lib.replaceContent(mdContent, "|:------|:------|:------------|\n", "\n\n[Top](#gepper-readme)", cMdGitIgnoreTableRows);
+    return mdContent;
 };
 var newContent;
 newContent = buildAndReplaceContent();
 console.log("==================\n"); console.log(newContent); console.log("==================\n");
 fs.writeFileSync(mdFile, newContent);
-console.log(`To create a new package with this \nversion, give the following command:\n\n    vsce package ${packageJson.version}\n`);
- 
+console.log(`Run optional commands to:\n`);
+console.log(` - To create a new package:`);
+console.log(`     vsce package ${packageJson.version}\n`);
+console.log(` - To install this new package for testing:`);
+console.log(`     code --install-extension gepper-${packageJson.version}.vsix\n\n`);
+