+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to the "gepper" extension will be documented in this file.
 ### Added
 - Command *Create new C++ CMake project with GoogleTest*. This command with create a C++ CMake boiler plate project, add one hello world library and one hello world application.
 Included are CTest and GoogleTest examples.
+- Added .gitignore specific snippets for *Visual Studio Code*, C*++* and *CMake*.
   
 -